+++ remoc/tests/chmux/mod.rs
@@ -1,4 +1,5 @@
 mod channel;
+mod global_credits;
 mod storage;
 
 #[cfg(not(target_family = "wasm"))]